martysiuk/compiler/requirements.txt
antlr4-python3-runtime>=4.9.0

# Опционально: для автоматической генерации парсера из грамматики
# antlr4-tools>=0.2.0

